+++ Classes/Model/BasePDF.php
@@ -47,23 +47,122 @@ class BasePDF extends FPDI
     protected $importTemplateOnThisPage = false;
 
     /**
+     * Document wide HeaderViewHelper
+     *
      * @var Closure
      */
-    protected $headerClosure = null;
+    protected $documentHeaderClosure = null;
 
     /**
+     * Page wide HeaderViewHelper, overwriting Document wide HeaderViewHelper
+     *
      * @var Closure
      */
-    protected $footerClosure = null;
+    protected $pageHeaderClosure = null;
+
+    /**
+     * @var string
+     */
+    protected $pageHeaderScope = null;
+
+    /**
+     * Document wide FooterViewHelper
+     *
+     * @var Closure
+     */
+    protected $documentFooterClosure = null;
+
+    /**
+     * Page wide FooterViewHelper, overwriting Document wide FooterViewHelper
+     *
+     * @var Closure
+     */
+    protected $pageFooterClosure = null;
+
+    /**
+     * Indicating whether page break is triggered by PageViewHelper or an auto page break
+     *
+     * @var boolean
+     */
+    protected $isAutoPageBreak = false;
+
+    /**
+     * @var string
+     */
+    protected $pageFooterScope = null;
+
+    const SCOPE_THIS_PAGE = 'thisPage';
+    const SCOPE_THIS_PAGE_INCLUDING_PAGE_BREAKS = 'thisPageIncludingPageBreaks';
+    const SCOPE_DOCUMENT = 'document';
+
+    /**
+     * @return void
+     */
+    protected function setHeader()
+    {
+        //disable TCPDF default behaviour in order to be able to overwrite header and footer on page level
+    }
+
+    /**
+     * @return void
+     */
+    protected function setFooter()
+    {
+        //disable TCPDF default behaviour in order to be able to overwrite header and footer on page level
+    }
+
+    /**
+     * Custom method to call setHeader to render header after the children of a PageViewHelper are rendered.
+     * This allows to overwrite header within a PageViewHelper.
+     *
+     * @return void
+     */
+    public function renderHeader()
+    {
+        parent::setHeader();
+    }
+
+    /**
+     * Custom method to call setFooter to render footer after the children of a PageViewHelper are rendered.
+     * This allows to overwrite footer within a PageViewHelper.
+     *
+     * @return void
+     */
+    public function renderFooter()
+    {
+        parent::setFooter();
+    }
+
+    /**
+     * Use this method uf you extend BasePDF and want to render a header by accessing the TCPDF API directly
+     *
+     * @return void
+     */
+    public function basePdfHeader()
+    {
+    }
+
+    /**
+     * Use this method uf you extend BasePDF and want to render a footer by accessing the TCPDF API directly
+     *
+     * @return void
+     */
+    public function basePdfFooter()
+    {
+    }
 
     /**
      * @return void
      */
     public function Header() // phpcs:ignore
     {
-        if ($this->headerClosure instanceof Closure) {
-            $this->headerClosure->__invoke();
+        if ($this->pageHeaderClosure instanceof Closure) {
+            $this->pageHeaderClosure->__invoke();
+        } else if ($this->documentHeaderClosure instanceof Closure) {
+            $this->documentHeaderClosure->__invoke();
         }
+
+        $this->basePdfHeader();
     }
 
     /**
@@ -71,9 +170,13 @@ class BasePDF extends FPDI
      */
     public function Footer() // phpcs:ignore
     {
-        if ($this->footerClosure instanceof Closure) {
-            $this->footerClosure->__invoke();
+        if ($this->pageFooterClosure instanceof Closure) {
+            $this->pageFooterClosure->__invoke();
+        } else if ($this->documentFooterClosure instanceof Closure) {
+            $this->documentFooterClosure->__invoke();
         }
+
+        $this->basePdfFooter();
     }
 
     /**
@@ -84,6 +187,19 @@ class BasePDF extends FPDI
     public function AddPage($orientation = '', $format = '', $rotationOrKeepmargins = false, $tocpage = false) // phpcs:ignore
     {
         parent::AddPage($orientation, $format, $rotationOrKeepmargins, $tocpage);
+
+        if ($this->isAutoPageBreak) {
+            if ($this->pageHeaderScope === self::SCOPE_THIS_PAGE) {
+                $this->pageHeaderClosure = null; //remove page header closure as it should only be used once
+            }
+
+            if ($this->pageFooterScope === self::SCOPE_THIS_PAGE) {
+                $this->pageFooterClosure = null; //remove page footer closure as it should only be used once
+            }
+
+            $this->renderHeader();
+            $this->renderFooter();
+        }
 
         if ($this->importTemplateOnThisPage && $this->tpl !== 0) {
             $this->useTemplate($this->tpl);
@@ -101,18 +217,42 @@ class BasePDF extends FPDI
     }
 
     /**
+     * @param Closure $closure
+     * @param string $scope
+     *
      * @return void
      */
-    public function setHeaderClosure(Closure $closure)
+    public function setHeaderClosure(Closure $closure = null, $scope = self::SCOPE_DOCUMENT)
     {
-        $this->headerClosure = $closure;
+        if ($scope === self::SCOPE_DOCUMENT) {
+            $this->documentHeaderClosure = $closure;
+        } else if (in_array($scope, [self::SCOPE_THIS_PAGE, self::SCOPE_THIS_PAGE_INCLUDING_PAGE_BREAKS])) {
+            $this->pageHeaderScope = $scope;
+            $this->pageHeaderClosure = $closure;
+        }
     }
 
     /**
+     * @param Closure $closure
+     * @param string $scope
+     *
      * @return void
      */
-    public function setFooterClosure(Closure $closure)
+    public function setFooterClosure(Closure $closure = null, $scope = self::SCOPE_DOCUMENT)
     {
-        $this->footerClosure = $closure;
+        if ($scope === self::SCOPE_DOCUMENT) {
+            $this->documentFooterClosure = $closure;
+        } else if (in_array($scope, [self::SCOPE_THIS_PAGE, self::SCOPE_THIS_PAGE_INCLUDING_PAGE_BREAKS])) {
+            $this->pageFooterScope = $scope;
+            $this->pageFooterClosure = $closure;
+        }
+    }
+
+    /**
+     * @param boolean $isAutoPageBreak
+     */
+    public function setIsAutoPageBreak($isAutoPageBreak)
+    {
+        $this->isAutoPageBreak = $isAutoPageBreak;
     }
 }